Create ReputationHandler.hpp
