+++ src/world/Units/Players/ReputationHandler.Legacy.cpp
@@ -26,17 +26,7 @@
 #include "Server/World.Legacy.h"
 #include "Management/Group.h"
 #include "Objects/ObjectMgr.h"
-
-enum FactionFlags
-{
-    FACTION_FLAG_VISIBLE            = 0x01,
-    FACTION_FLAG_AT_WAR             = 0x02,
-    FACTION_FLAG_HIDDEN             = 0x04,
-    FACTION_FLAG_FORCED_INVISIBLE   = 0x08,     // if both ACTION_FLAG_VISIBLE and FACTION_FLAG_FORCED_INVISIBLE are set, client crashes!
-    FACTION_FLAG_DISABLE_ATWAR      = 0x10,     // disables AtWar button for client, but you can be in war with the faction
-    FACTION_FLAG_INACTIVE           = 0x20,
-    FACTION_FLAG_RIVAL              = 0x40      // only Scryers and Aldor have this flag
-};
+#include "ReputationHandler.hpp"
 
 Standing Player::GetReputationRankFromStanding(int32 Standing_)
 {
